requirements.txt
# Core async HTTP and dependencies
httpx>=0.24.0
tenacity>=8.2.0
aiosqlite>=0.19.0

# Data parsing and environment
xmltodict>=0.13.0
python-dotenv>=1.0.0
pydantic>=2.0.0

# LLM for disease-to-target mapping
openai>=0.28.0

# Additional dependencies
tabulate>=0.9.0

# Optional LangChain integration (uncomment if using LLM orchestration)
# langchain>=0.1.0
# langchain-core>=0.1.0
# langchain-community>=0.0.10
# langchain-openai>=0.0.5
